!catreset doesn't delete active shell channels anymore

+++ shellman/frontends/discord_frontend/main_channel_handlers/reset_category.py
@@ -6,6 +6,12 @@ async def handle(message, match, discord_client):
         await message.channel.send('Cannot reset category in non-admin mode.')
         return
     for channel in discord_client.category.channels:
+        active = None
+        for index, shell in discord_client.shellman_frontend.shells.items():
+            if active := shell.channel == channel:
+                break
+        if active:
+            continue
         await channel.delete()
